Время упр, статистика, ккалории ко всем упражнениям, уменьшение и увеличение ккалорий при усложнении тренировки

+++ app/src/main/java/com/example/fitnessapp/statistic/data/DaysFinishStateList.kt
@@ -3,7 +3,7 @@ package com.example.fitnessapp.statistic.data
 data class DaysFinishStateList(
     var trainingCounter : Int = 0,
     var date: String, //Будем показывать дату
-    var kcal: Int, // Сколько всего сжег ккалорий
+    var kcal: Double, // Сколько всего сжег ккалорий
     var workoutTime: String // Будем показывать сколько всего он занимался
 
 )
